% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addmonths.R
\name{addmonths}
\alias{addmonths}
\title{Add a specified number of months to a date}
\usage{
addmonths(dt, m)
}
\arguments{
\item{dt}{A date or vector of dates.}

\item{m}{An integer or vector of integers representing the number of months to add.}
}
\value{
A date or vector of dates with the number of months added.
}
\description{
Adds a specified number of months to a given date.  If the number of days in the start date is greater than the number of days in the month of the end date, the last day of the month is returned.
}
\examples{
addmonths(as.Date("2015-01-31"),13)
addmonths(as.Date("2015-02-28"),-2)
addmonths(as.Date(c("2015-01-31","2015-02-28")),-2)
addmonths(as.Date(c("2015-01-31","2015-02-28")),c(-1,-2))
}
