% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daysinmonth.R
\name{daysinmonth}
\alias{daysinmonth}
\title{Number of days in a month}
\usage{
daysinmonth(dt)
}
\arguments{
\item{dt}{A date or vector of dates.}
}
\value{
The number of days in the month of a given date.
}
\description{
Calculates the number of days in the month of a given date.
}
\examples{
daysinmonth(as.Date("2016-01-15"))
daysinmonth(as.Date(c("2015-02-19","2016-02-11")))
}
