% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modiff5.R
\name{modiff5}
\alias{modiff5}
\title{Approximate number of months between two dates rounded to nearest half month}
\usage{
modiff5(dt1, dt2)
}
\arguments{
\item{dt1}{The starting date or vector of dates.}

\item{dt2}{The ending date or vector of dates.}
}
\value{
The approximate number of months between the dates rounded to the nearest half month.
}
\description{
Calculates the approximate number of months between two dates rounded to the nearest half month.  This function was created with the idea that dt1 and dt2 are already rounded using rnddt().
}
\examples{
modiff5(as.Date("2015-01-01"),as.Date("2015-02-23"))
modiff5(as.Date("2015-01-01"),as.Date("2015-02-24"))
}
