% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olf.R
\name{olf}
\alias{olf}
\title{Calculate on-level factors}
\usage{
olf(in.rc, in.yr, in.prem.type, in.year.type, in.policy.length)
}
\arguments{
\item{in.rc}{The rate change history.  in.rc is a data frame with three columns.  Column 1 is named "effdt" and is of type Date.  Column 2 is named "rc" and is of type numeric (note that rc is centered around zero, and not a rate level factor centered around unity).  Column 3 is named "type" and is of type char.  Possible values are either "R" for rate change or "L" for law change.}

\item{in.yr}{The historical periods for which the on-level factors will be calculated.  in.yr is a data frame with two columns.  Column 1 is named "effdt" and if of type Date.  effdt is the inclusive effective date.  Column 2 is named "expdt" and is of type Date.  expdt is the exclusive expiration date.}

\item{in.prem.type}{"E" = earned premium, "W" = written premium}

\item{in.year.type}{"C" = calendar year, "P" = policy year}

\item{in.policy.length}{Integer representing the length of a policy in months.}
}
\value{
A list of five on-level factor components.
}
\description{
Calculate on-level factors
}
