\docType{package}
\name{OnLevel-package}
\alias{OnLevel}
\alias{OnLevel-package}
\title{The OnLevel Package}
\description{
  On Level.
}
\details{
  \tabular{ll}{ Package: \tab OnLevel\cr Type: \tab
  Package\cr Version: \tab 0.01-1\cr Date: \tab
  2013-06-17\cr Depends: \tab R (>= 2.11.0)\cr Encoding:
  \tab UTF-8\cr License: \tab GPL (>= 3)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Brian Fannin \email{BFannin@RedwoodsGroup.com}
}
\keyword{package}

