% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnddt.R
\name{rnddt}
\alias{rnddt}
\title{Round a date to the nearest half month}
\usage{
rnddt(dt)
}
\arguments{
\item{dt}{A date or vector of dates to be rounded.}
}
\value{
The rounded date or vector of dates.
}
\description{
Takes a date or vector of dates and rounds each value to the nearest half month.  Days 1 to 8 get rounded down to 1, days 9 to 23 get rounded to the 15th, and days 23+ get rounded to the 1st of the following month.
}
\examples{
dt <- seq(as.Date("2015-01-01"), to = as.Date("2015-01-31"), by = "day")
data.frame(dt = dt, rnddt = rnddt(dt))
}
